requirements/devel.txt
-r base.txt
django-celery
pysolr
psycopg2